rename session to token, modify gorm config, remove TableName in IPaste interface
