+++ model/paste/paste.go
@@ -11,7 +11,6 @@ import (
 type IPaste interface {
 	Save() error
 	Get(string) error
-	TableName() string
 	Delete() error
 	GetKey() string
 	GetContent() string
